// app/enums/page-types.enum.ts
export enum PageTypes {
    brandLanding = 'brand-landing-page',
    browse = 'browse-page',
    category = 'category-page',
    content = 'contentpage',
    productDetails = 'product-details-page',
    pageNotFound = 'pagenotfound',
    home = 'homepage',
    whereToBuy = 'find-a-store-basic-page',
    proPlusLanding = 'proplus-landing-page',
}